Jackson - Centralize recording in a unique recorder
I will have to add some recording and I don't think creating one more
specialized Recorder is the way to go.
Also added the phase annotations for documentation.
