+++ extensions/jackson/runtime/src/main/java/io/quarkus/jackson/runtime/JacksonRecorder.java
@@ -1,15 +1,20 @@
 package io.quarkus.jackson.runtime;
 
+import java.util.Map;
 import java.util.Optional;
 import java.util.function.Supplier;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 
+import io.quarkus.jackson.ObjectMapperCustomizer;
 import io.quarkus.runtime.annotations.Recorder;
+import io.quarkus.runtime.annotations.StaticInit;
 
 @Recorder
-public class JacksonSupportRecorder {
+public class JacksonRecorder {
 
+    @StaticInit
     public Supplier<JacksonSupport> supplier(Optional<String> propertyNamingStrategyClassName) {
         return new Supplier<>() {
             @Override
@@ -31,6 +36,28 @@ public class JacksonSupportRecorder {
                             }
                         }
                         return Optional.empty();
+                    }
+                };
+            }
+        };
+    }
+
+    @StaticInit
+    public Supplier<ObjectMapperCustomizer> customizerSupplier(Map<Class<?>, Class<?>> mixinsMap) {
+        return new Supplier<>() {
+            @Override
+            public ObjectMapperCustomizer get() {
+                return new ObjectMapperCustomizer() {
+                    @Override
+                    public void customize(ObjectMapper objectMapper) {
+                        for (var entry : mixinsMap.entrySet()) {
+                            objectMapper.addMixIn(entry.getKey(), entry.getValue());
+                        }
+                    }
+
+                    @Override
+                    public int priority() {
+                        return DEFAULT_PRIORITY + 1;
                     }
                 };
             }
